add scaladoc and rename some methods

+++ app/services/MemberService.scala
@@ -3,7 +3,6 @@ package services
 import javax.inject.Inject
 import play.api.db.slick.{DatabaseConfigProvider, HasDatabaseConfigProvider}
 import formats.{MemberCommand, MemberView}
-import play.api.libs.json.JsValue
 import play.api.mvc.{AnyContent, Request}
 import repositories.MemberRepositoryJDBC
 import models.Tables.Member
@@ -13,7 +12,7 @@ import scala.concurrent.{ExecutionContext, Future}
 import slick.driver.JdbcProfile
 
 /**
-  * Memberテーブルに対してDB接続をするクラス
+  * Memberテーブルに関するビジネスロジックを実装するクラス
   *
   * @author yuito.sato
   */
@@ -26,10 +25,17 @@ class MemberService @Inject()(val memberJdbc: MemberRepositoryJDBC, val dbConfig
     } yield createMember
   }
 
+  /**
+    * サインイン中の会員のIDを取得するメソッド。
+    * この返り値がSomeかどうかでサインインステータスを確認することも可能
+    *
+    * @param rs クライアントからのリクエスト
+    * @return キャッシュから取得したMemberテーブルのmemberId
+    */
   def findCurrentMemberId(implicit rs: Request[AnyContent]): Option[Long] = {
     for {
       token <- rs.session.get("token")
-      memberId <- cache.get[Int](token)
+      memberId <- cache.get[Long](token)
     } yield memberId
   }
